+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Won't prompt for name if `--yes` flag is given.
+
 ## [v0.21.0](https://github.com/allenai/beaker-gantry/releases/tag/v0.21.0) - 2024-01-30
 
 ### Added