+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.22.0](https://github.com/allenai/beaker-gantry/releases/tag/v0.22.0) - 2024-02-28
+
 ### Added
 
 - Added flag `--stop-preemptible` to `gantry run` for stopping all preemptible jobs on the cluster. This requires that your job is not preemptible and you've specified a single cluster.